# Use SSE2 in the x86_64 C version of decodeLatin1