Before (sorry for truncated lines):

```
benchmarking DecodeASCII/strict decodeLatin1
time                 28.60 ns   (28.53 ns .. 28.67
                     1.000 R²   (1.000 R² .. 1.000
mean                 28.60 ns   (28.50 ns .. 28.69
std dev              324.4 ps   (258.4 ps .. 426.7
variance introduced by outliers: 12% (moderately i

benchmarking DecodeASCII/lazy decodeLatin1
time                 39.73 ns   (39.42 ns .. 40.07
                     1.000 R²   (1.000 R² .. 1.000
mean                 39.56 ns   (39.46 ns .. 39.74
std dev              431.1 ps   (271.5 ps .. 699.6
variance introduced by outliers: 11% (moderately i
```

After:

```
benchmarking DecodeASCII/strict decodeLatin1
time                 21.71 ns   (21.67 ns .. 21.77 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 21.70 ns   (21.67 ns .. 21.74 ns)
std dev              114.6 ps   (90.46 ps .. 150.4 ps)

benchmarking DecodeASCII/lazy decodeLatin1
time                 32.80 ns   (32.75 ns .. 32.84 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 32.89 ns   (32.78 ns .. 33.39 ns)
std dev              660.6 ps   (137.1 ps .. 1.483 ns)
variance introduced by outliers: 29% (moderately inflated)
```